+++ packages/perseus-core/src/index.ts
@@ -140,6 +140,7 @@ export {
 } from "./widgets/apply-defaults";
 
 export {default as getDefaultAnswerArea} from "./utils/get-default-answer-area";
+export {getDefaultFigureForType} from "./utils/get-default-figure-for-type";
 export {
     default as splitPerseusItem,
     splitPerseusItemJSON,
@@ -208,9 +209,17 @@ export {
 
 // Generators
 export {
+    generateDefinitionOptions,
+    generateDefinitionWidget,
+} from "./utils/generators/definition-widget-generator";
+export {
     generateDropdownOptions,
     generateDropdownWidget,
 } from "./utils/generators/dropdown-widget-generator";
+export {
+    generateExplanationOptions,
+    generateExplanationWidget,
+} from "./utils/generators/explanation-widget-generator";
 export {
     generateExpressionOptions,
     generateExpressionAnswerForm,