+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/locked-function-settings.test.tsx
@@ -1,3 +1,4 @@
+import {getDefaultFigureForType} from "@khanacademy/perseus-core";
 import {RenderStateRoot} from "@khanacademy/wonder-blocks-core";
 // eslint-disable-next-line testing-library/no-manual-cleanup
 import {render, screen, cleanup} from "@testing-library/react";
@@ -5,10 +6,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
 import LockedFunctionSettings from "./locked-function-settings";
-import {
-    getDefaultFigureForType,
-    mockedJoinLabelsAsSpokenMathForTests,
-} from "./util";
+import {mockedJoinLabelsAsSpokenMathForTests} from "./util";
 
 import type {Props} from "./locked-function-settings";
 import type {UserEvent} from "@testing-library/user-event";