+++ packages/perseus/src/__testdata__/renderer.testdata.ts
@@ -1,6 +1,9 @@
 import {
+    generateDefinitionOptions,
+    generateDefinitionWidget,
     generateDropdownOptions,
     generateDropdownWidget,
+    generateTestPerseusRenderer,
 } from "@khanacademy/perseus-core";
 
 import type {MockWidget} from "../widgets/mock-widgets/mock-widget-types";
@@ -77,49 +80,30 @@ export const question2: PerseusRenderer = {
     widgets: {"mock-widget 1": mockWidget},
 };
 
-export const definitionItem: PerseusRenderer = {
+export const definitionItem: PerseusRenderer = generateTestPerseusRenderer({
     content:
         "Mock widgets ==> [[\u2603 definition 1]] [[\u2603 definition 2]] [[\u2603 definition 3]]",
-    images: {},
     widgets: {
-        "definition 1": {
-            graded: true,
-            version: {major: 0, minor: 0},
-            static: false,
-            type: "definition",
-            options: {
+        "definition 1": generateDefinitionWidget({
+            options: generateDefinitionOptions({
                 togglePrompt: "word",
                 definition: "",
-                static: false,
-            },
-            alignment: "default",
-        },
-        "definition 2": {
-            graded: true,
-            version: {major: 0, minor: 0},
-            static: false,
-            type: "definition",
-            options: {
+            }),
+        }),
+        "definition 2": generateDefinitionWidget({
+            options: generateDefinitionOptions({
                 togglePrompt: "word",
                 definition: "",
-                static: false,
-            },
-            alignment: "default",
-        },
-        "definition 3": {
-            graded: true,
-            version: {major: 0, minor: 0},
-            static: false,
-            type: "definition",
-            options: {
+            }),
+        }),
+        "definition 3": generateDefinitionWidget({
+            options: generateDefinitionOptions({
                 togglePrompt: "word",
                 definition: "",
-                static: false,
-            },
-            alignment: "default",
-        },
+            }),
+        }),
     },
-} as PerseusRenderer;
+});
 
 export const mockedRandomItem: PerseusRenderer = {
     content: "Mock widgets ==> [[\u2603 radio 1]] [[\u2603 radio 2]]",